sockets: Make the server really work with udp via using a cache system
This creates a vector of data, which we look-up to retrieve the
connection state.

Yes, linear search again, and more than probably subject to IP spoofing
and a lot of other things, but it's kind of nice that it works out of
the box...

The solution proposed by our teacher was to create a new ephimerous
socket instead, and start another thread to manage it just like the tcp
one... I thought it was really under-performant.

Anyway, it works now, which is the important part :)